Add support for mac commands sending to end device

+++ models/packets.go
@@ -63,7 +63,7 @@ type GatewayStatsPacket struct {
 type RXPayload struct {
 	DevEUI       lorawan.EUI64 `json:"devEUI"`
 	FPort        uint8         `json:"fPort"`
-	GatewayCount int           `json:"gatewayCount"`
+	GatewayCount uint8         `json:"gatewayCount"`
 	RSSI         int           `json:"rssi"`
 	Data         []byte        `json:"data"`
 }
